Fixed error in displaying the percentage of completion

+++ math.go
@@ -1,5 +1,5 @@
 package main
 
-func percentageOfNumber(first int, second int) int {
+func percentageOfNumber(first float64, second float64) float64 {
 	return (first / second) * 100
 }
